foutje in events

+++ ChipDistributionGame/Shared/src/main/java/nl/han/asdb/shared/interfaces/eventmanager/events/game/OrderChipsEvent.java
@@ -47,7 +47,7 @@ public class OrderChipsEvent extends Event {
         return amount;
     }
 
-    public void setAmount(short amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
